samples/custom-operation/requirements.txt
accelbyte-py-sdk
bitarray
httpx
mmh3
PyJWT[crypto]
python-dotenv
PyYAML
requests
websockets